+++ src/aioverse/models/contexts/assistant_context.py
@@ -5,4 +5,7 @@ from typing	import Literal
 
 
 class AssistantContext(BaseContext):
+
+	"""Assistant 类型上下文"""
+
 	role: Literal[Roles.ASSISTANT] = Roles.ASSISTANT